+++ packages/cheqd-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/cheqd-blockchain-modules
 
+## 0.2.0
+
+### Minor Changes
+
+- Change `DIDDocument` modification behaviour to throw error on attempt to add key/service/controller with duplicate id.
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.2.0
+
 ## 0.1.1
 
 ### Minor Changes