+++ packages/credential-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/credential-sdk
 
+## 0.2.0
+
+### Minor Changes
+
+- Change `DIDDocument` modification behaviour to throw error on attempt to add key/service/controller with duplicate id.
+
 ## 0.1.1
 
 ### Minor Changes